fix(tmpx): address golangci-lint findings
- Split the const block in tmpx.go so the typed TmpxFormatVersion isn't
  hiding an untyped tmpxKidMaxLen alongside it (SA9004).
- Annotate the count-byte conversion with //nolint:gosec — len(entries)
  is bounds-checked to ≤255 immediately above (G115).
- Convert through int instead of byte in the test assertion (G115).

+++ tmproto/tmpx_test.go
@@ -224,7 +224,7 @@ func TestSealTmpxRoundtrip(t *testing.T) {
 	if string(got[5:7]) != "US" {
 		t.Errorf("country: got %q, want US", got[5:7])
 	}
-	if got[15] != byte(len(entries)) {
+	if int(got[15]) != len(entries) {
 		t.Errorf("count: got %d, want %d", got[15], len(entries))
 	}
 }
